backend/requirements.txt
#============= Common =============#
# py39
celery==5.6.3
redis==7.0.1

# py312
# celery==5.6.2
# redis==7.4.0

#============= app.py =============#

# py312
# keras==3.14.0
# matplotlib==3.10.8
# tensorflow==2.21.0

# py39
opencv-python==4.13.0.92
keras==3.10.0
matplotlib==3.9.4
pandas==2.3.3
tensorflow==2.20.0
pyspark==3.5.0

#============= Django =============#
# py312
# Django==6.0.2
# djangorestframework==3.17.0
# mysqlclient==2.2.8


# py39
Django==4.2.30
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
# django-phone-verify==3.3.0
django-redis==6.0.0
mysqlclient==2.2.7


#============= FastAPI =============#
# py312
# sqlalchemy==2.0.48
# fastapi==0.128.1
# httpx==0.25.2
# langchain==1.2.13
# langchain-openai==0.0.5
# pydantic-settings==2.1.0
# uvicorn[standard]==0.24.0


# py39
sqlalchemy==2.0.49
motor==3.7.1
python-jose[cryptography]==3.5.0

fastapi==0.128.8
uvicorn==0.39.0
asyncmy==0.2.11
pika==1.3.2
httpx==0.28.1
langchain==0.3.28
langchain-openai==0.3.35
pydantic-settings==2.11.0



